Fix wrong colors in ThemeOptionScreen
