+++ feature/theme/build.gradle
@@ -29,7 +29,6 @@ dependencies {
     implementation Libs.Compose.foundation
     implementation Libs.Compose.material
     implementation Libs.Compose.ui
-    implementation Libs.Compose.runtime_livedata
     implementation Libs.Accompanist.insets
 
     implementation Libs.AndroidX.Navigation.fragment
